% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclass.R
\name{MyClass}
\alias{MyClass}
\title{Create an instance of MyClass}
\value{
An instance of MyClass
}
\description{
Create an instance of MyClass

Create an instance of MyClass
}
\examples{
\dontshow{if (TRUE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
MyClass$new()$bar()
\dontshow{\}) # examplesIf}

## ------------------------------------------------
## Method `MyClass$foo`
## ------------------------------------------------

MyClass$new()$foo()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MyClass-new}{\code{MyClass$new()}}
\item \href{#method-MyClass-foo}{\code{MyClass$foo()}}
\item \href{#method-MyClass-bar}{\code{MyClass$bar()}}
\item \href{#method-MyClass-clone}{\code{MyClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MyClass} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-foo"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-foo}{}}}
\subsection{Method \code{foo()}}{
Do something.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$foo()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{MyClass$new()$foo()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-bar"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-bar}{}}}
\subsection{Method \code{bar()}}{
Do something else.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$bar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
